// src/app/model/gauges-data-list.ts
export interface GaugesDataList {
  noDataPolicyFound: string;
  plate: string;
  ownerName: string;
  notificationId: string;
  expert: string;
  notification: string;
  accidentTown: string;
  operator: string;
  natureCode: string;
  natureDesc: string;
  brandTrademark: string;
  nature: string;
  towingCom: string;
  insCompany: string;
  expertDispDate: string;
  reportedDateTime: string;
  noDataType: string;
  towingTo: string;
  towingDispDate: string;
  towingFrom: string;
}